% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_circletype.R
\name{use_circletype}
\alias{use_circletype}
\title{Load CircleType dependencies}
\usage{
use_circletype()
}
\value{
Load depepdencies
}
\description{
Load CircleType dependencies
}
\examples{
if (interative()) {
library(shiny)

ui <- fluidPage(
  use_circletype(), # load dependencies
  circletype(h1("this is a title", id = "test_h1"), dir = -1, radius = 200)
)

server <- function(input, output, session) {}

shinyApp(ui, server)
}
}
